# Conflicts:
#	charts-enterprise-modules/ag-charts-enterprise/README.md
#	charts-enterprise-modules/ag-charts-enterprise/gulpfile.js
#	charts-enterprise-modules/ag-charts-enterprise/package.json
#	charts-enterprise-modules/ag-charts-enterprise/src/animation/animation.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/animation/animationModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/background/background.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/background/backgroundImage.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/background/backgroundModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/context-menu/contextMenu.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/context-menu/contextMenuModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/context-menu/placeholder.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/crosshair/__image_snapshots__/crosshair-test-ts-crosshair-should-snap-to-the-closest-highlighted-node-datum-6-snap.png
#	charts-enterprise-modules/ag-charts-enterprise/src/crosshair/crosshair.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/crosshair/crosshair.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/crosshair/crosshairModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/gradient-legend/gradientLegend.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/gradient-legend/gradientLegend.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/gradient-legend/main.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/heatmap/heatmap.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/heatmap/heatmapDefaults.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/heatmap/heatmapModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/heatmap/heatmapSeries.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/heatmap/typings.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/license/licenseManager.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/main.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/navigator/chartNavigatorModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/navigator/navigator.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/navigator/navigatorHandle.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/navigator/navigatorMask.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/navigator/placeholder.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/polar-series/radar-line/main.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/__image_snapshots__/waterfall-test-ts-chart-should-render-a-waterfall-bar-chart-as-expected-1-snap.png
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/__image_snapshots__/waterfall-test-ts-chart-should-render-a-waterfall-column-chart-as-expected-1-snap.png
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/typings.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/waterfall.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/waterfallDefaults.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/waterfallModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/waterfallSeries.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/waterfall/waterfallThemes.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoom.test.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoom.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoomModule.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoomPanner.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoomScroller.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoomSelector.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoomTransformers.ts
#	charts-enterprise-modules/ag-charts-enterprise/src/zoom/zoomTypes.ts
#	charts-enterprise-modules/ag-charts-enterprise/tsconfig.cjs.es5.json
#	charts-enterprise-modules/ag-charts-enterprise/tsconfig.cjs.es6.json
#	charts-enterprise-modules/ag-charts-enterprise/tsconfig.esm.es5.json
#	charts-enterprise-modules/ag-charts-enterprise/tsconfig.esm.es6.json
#	grid-community-modules/react/lib/reactUi/jsComp.js
#	grid-community-modules/react/lib/reactUi/useEffectOnce.js
#	grid-packages/ag-grid-react/lib/reactUi/jsComp.js
#	grid-packages/ag-grid-react/lib/reactUi/useEffectOnce.js

+++ grid-enterprise-modules/server-side-row-model/src/serverSideRowModel/stores/lazy/lazyCache.ts
@@ -1,4 +1,4 @@
-import { Autowired, BeanStub, FocusService, GridApi, LoadSuccessParams, NumberSequence, PostConstruct, PreDestroy, RowNode, IRowNode, ServerSideGroupLevelParams, GetRowIdFunc, WithoutGridCommon, GetRowIdParams } from "@ag-grid-community/core";
+import { Autowired, BeanStub, FocusService, GridApi, LoadSuccessParams, NumberSequence, PostConstruct, PreDestroy, RowNode, IRowNode, ServerSideGroupLevelParams, GetRowIdFunc, WithoutGridCommon, GetRowIdParams, Events } from "@ag-grid-community/core";
 import { BlockUtils } from "../../blocks/blockUtils";
 import { NodeManager } from "../../nodeManager";
 import { LazyStore } from "./lazyStore";
@@ -60,6 +60,13 @@ export class LazyCache extends BeanStub {
      */
     private getRowIdFunc?: ((params: WithoutGridCommon<GetRowIdParams<any, any>>) => string);
     private isMasterDetail: boolean;
+
+    /**
+     * A cache of removed group nodes, this is retained for preserving group
+     * state when the node moves in and out of the cache. Generally caused by
+     * rows moving blocks.
+     */
+    private removedNodeCache = new Map<string, RowNode>();
 
     constructor(store: LazyStore, numberOfRows: number, storeParams: ServerSideGroupLevelParams) {
         super();
@@ -392,6 +399,20 @@ export class LazyCache extends BeanStub {
         if (data && this.getRowIdFunc != null) {
             const id = this.getRowId(data);
 
+            // the node was deleted at some point, but as we're refreshing
+            // it's been cached and we can retrieve it for reuse.
+            const deletedNode = id && this.removedNodeCache?.get(id);
+            if (deletedNode) {
+                this.removedNodeCache?.delete(id!);
+                this.blockUtils.updateDataIntoRowNode(deletedNode, data);
+                this.nodeMap.set({
+                    id: deletedNode.id!,
+                    node: deletedNode,
+                    index: atStoreIndex
+                });
+                return deletedNode;
+            }
+
             const lazyNode = this.nodeMap.getBy('id', id);
             if (lazyNode) {
                 // delete old lazy node so we can insert it at different location
@@ -404,6 +425,13 @@ export class LazyCache extends BeanStub {
                     node,
                     index: atStoreIndex
                 });
+                this.nodesToRefresh.delete(node);
+
+                if (this.rowLoader.getBlockStartIndexForIndex(index) === this.rowLoader.getBlockStartIndexForIndex(atStoreIndex)) {
+                    // if the block hasn't changed and we have a nodes map, we don't need to refresh the original block, as this block
+                    // has just been refreshed.
+                    return node;
+                }
 
                 // mark all of the old block as needsVerify to trigger it for a refresh, as nodes
                 // should not be out of place
@@ -499,7 +527,13 @@ export class LazyCache extends BeanStub {
         this.nodeDisplayIndexMap.delete(lazyNode.node.rowIndex!);
         this.nodesToRefresh.delete(lazyNode.node);
         
-        this.blockUtils.destroyRowNode(lazyNode.node);
+        if (lazyNode.node.group && this.nodesToRefresh.size > 0) {
+            // while refreshing, we retain the group nodes so they can be moved
+            // without losing state
+            this.removedNodeCache.set(lazyNode.node.id!, lazyNode.node);
+        } else {
+            this.blockUtils.destroyRowNode(lazyNode.node);
+        }
     }
 
     public getSsrmParams() {
@@ -717,6 +751,7 @@ export class LazyCache extends BeanStub {
                 return;
             }
 
+            // node already exists, and same as node at designated position, update data
             if (nodeFromCache && this.doesNodeMatch(data, nodeFromCache.node)) {
                 this.blockUtils.updateDataIntoRowNode(nodeFromCache.node, data);
                 this.nodesToRefresh.delete(nodeFromCache.node);
@@ -763,6 +798,13 @@ export class LazyCache extends BeanStub {
         if (!finishedRefreshing) {
             return;
         }
+
+        // any nodes left in the map need to be cleaned up, this prevents us preserving nodes
+        // indefinitely
+        this.removedNodeCache.forEach(node => {
+            this.blockUtils.destroyRowNode(node);
+        });
+        this.removedNodeCache = new Map();
 
         this.store.fireRefreshFinishedEvent();
     }
